refactor: simplify code
adding unit tests for script engine
