+++ options.go
@@ -1,0 +1,29 @@
+package jsonpath
+
+import (
+	"github.com/evilmonkeyinc/jsonpath/script"
+)
+
+// OptionFunction function that can be used as a compile or query option
+type OptionFunction func(selector *Selector) error
+
+// Apply the option to the Selector
+func (fn OptionFunction) Apply(selector *Selector) error {
+	return fn(selector)
+}
+
+// Option allows to set compile and query options when calling Compile
+type Option interface {
+	// Apply the option to the Selector
+	Apply(selector *Selector) error
+}
+
+// ScriptEngine allows you to set a custom script Engine for the JSONPath selector
+func ScriptEngine(engine script.Engine) Option {
+	return OptionFunction(func(selector *Selector) error {
+		if selector.engine == nil {
+			selector.engine = engine
+		}
+		return nil
+	})
+}
