+++ script/standard/engine.go
@@ -8,6 +8,8 @@ import (
 	"github.com/evilmonkeyinc/jsonpath/script"
 )
 
+// TODO : add tests for what is in readme
+// TODO : update readme to give more details, maybe add readme to this package and link from main
 // TODO : add support for bitwise operators | &^ ^ &  << >> after + and -
 var defaultTokens []string = []string{
 	"||", "&&",
@@ -37,6 +39,19 @@ func (engine *ScriptEngine) Compile(expression string, options *option.QueryOpti
 	}, nil
 }
 
+// Evaluate return the result of the expression evaluation
+func (engine *ScriptEngine) Evaluate(root, current interface{}, expression string, options *option.QueryOptions) (interface{}, error) {
+	compiled, err := engine.Compile(expression, options)
+	if err != nil {
+		return nil, err
+	}
+	evaluation, err := compiled.Evaluate(root, current)
+	if err != nil {
+		return nil, err
+	}
+	return evaluation, nil
+}
+
 func (engine *ScriptEngine) buildOperators(expression string, tokens []string, options *option.QueryOptions) (operator, error) {
 	nextToken := tokens[0]
 	expression = strings.TrimSpace(expression)
@@ -59,6 +74,7 @@ func (engine *ScriptEngine) buildOperators(expression string, tokens []string, o
 	}
 
 	// check right for more tokens, or use raw string as input
+	// right check needs done first as some expressions just have left sides
 	arg2Str := strings.TrimSpace(expression[idx+(len(nextToken)):])
 	if arg2Str == "" && (nextToken != "$" && nextToken != "@") {
 		if len(tokens) == 1 {
@@ -68,25 +84,9 @@ func (engine *ScriptEngine) buildOperators(expression string, tokens []string, o
 		return engine.buildOperators(expression, tokens[1:], options)
 	}
 
-	var arg2 interface{} = arg2Str
-	if op, err := engine.buildOperators(arg2Str, tokens, options); err != nil {
+	arg2, err := engine.parseArgument(arg2Str, tokens, options)
+	if err != nil {
 		return nil, err
-	} else if op != nil {
-		arg2 = op
-	} else {
-		arg2Str = strings.TrimSpace(arg2Str)
-		arg2 = arg2Str
-		if strings.HasPrefix(arg2Str, "[") && strings.HasSuffix(arg2Str, "]") {
-			val := make([]interface{}, 0)
-			if err := json.Unmarshal([]byte(arg2Str), &val); err == nil {
-				arg2 = val
-			}
-		} else if strings.HasPrefix(arg2Str, "{") && strings.HasSuffix(arg2Str, "}") {
-			val := make(map[string]interface{})
-			if err := json.Unmarshal([]byte(arg2Str), &val); err == nil {
-				arg2 = val
-			}
-		}
 	}
 
 	// check left for more tokens, or use raw string as input
@@ -99,30 +99,18 @@ func (engine *ScriptEngine) buildOperators(expression string, tokens []string, o
 		return engine.buildOperators(expression, tokens[1:], options)
 	}
 
-	var arg1 interface{} = arg1Str
-	if op, err := engine.buildOperators(arg1Str, tokens, options); err != nil {
+	arg1, err := engine.parseArgument(arg1Str, tokens, options)
+	if err != nil {
 		return nil, err
-	} else if op != nil {
-		arg1 = op
-	} else {
-		arg1Str = strings.TrimSpace(arg1Str)
-		arg1 = arg1Str
-		if strings.HasPrefix(arg1Str, "[") && strings.HasSuffix(arg1Str, "]") {
-			val := make([]interface{}, 0)
-			if err := json.Unmarshal([]byte(arg1Str), &val); err == nil {
-				arg1 = val
-			}
-		} else if strings.HasPrefix(arg1Str, "{") && strings.HasSuffix(arg1Str, "}") {
-			val := make(map[string]interface{})
-			if err := json.Unmarshal([]byte(arg1Str), &val); err == nil {
-				arg1 = val
-			}
-		}
 	}
 
 	switch nextToken {
 	case "@", "$":
-		return newSelectorOperator(expression, engine, options)
+		selector, err := newSelectorOperator(expression, engine, options)
+		if err != nil {
+			return nil, err
+		}
+		return selector, nil
 	case "=~":
 		return &regexOperator{
 			arg1: arg1,
@@ -205,52 +193,25 @@ func (engine *ScriptEngine) buildOperators(expression string, tokens []string, o
 	return nil, errUnsupportedOperator
 }
 
-// Evaluate return the result of the expression evaluation
-func (engine *ScriptEngine) Evaluate(root, current interface{}, expression string, options *option.QueryOptions) (interface{}, error) {
-	compiled, err := engine.Compile(expression, options)
-	if err != nil {
+func (engine *ScriptEngine) parseArgument(argument string, tokens []string, options *option.QueryOptions) (interface{}, error) {
+	if op, err := engine.buildOperators(argument, tokens, options); err != nil {
 		return nil, err
-	}
-	return compiled.Evaluate(root, current)
-}
-
-type compiledExpression struct {
-	expression   string
-	rootOperator operator
-	engine       *ScriptEngine
-	options      *option.QueryOptions
-}
-
-func (compiled *compiledExpression) Evaluate(root, current interface{}) (interface{}, error) {
-	expression := compiled.expression
-	if expression == "" {
-		return nil, getInvalidExpressionEmptyError()
-	}
-	parameters := map[string]interface{}{
-		"$":    root,
-		"@":    current,
-		"nil":  nil,
-		"null": nil,
+	} else if op != nil {
+		return op, nil
 	}
 
-	if compiled.rootOperator == nil {
-		if val, ok := parameters[expression]; ok {
-			return val, nil
+	argument = strings.TrimSpace(argument)
+	var arg interface{} = argument
+	if strings.HasPrefix(argument, "[") && strings.HasSuffix(argument, "]") {
+		val := make([]interface{}, 0)
+		if err := json.Unmarshal([]byte(argument), &val); err == nil {
+			arg = val
 		}
-
-		if number, err := getNumber(expression, parameters); err == nil {
-			return number, nil
-		} else if boolean, err := getBoolean(expression, parameters); err == nil {
-			return boolean, nil
+	} else if strings.HasPrefix(argument, "{") && strings.HasSuffix(argument, "}") {
+		val := make(map[string]interface{})
+		if err := json.Unmarshal([]byte(argument), &val); err == nil {
+			arg = val
 		}
-
-		return expression, nil
 	}
-
-	value, err := compiled.rootOperator.Evaluate(parameters)
-	if err != nil {
-		return nil, err
-	}
-
-	return value, nil
+	return arg, nil
 }
