+++ script/standard/engine_test.go
@@ -62,6 +62,14 @@ func Test_ScriptEngine_Compile(t *testing.T) {
 				},
 			},
 		},
+		{
+			input: input{
+				expression: ".$",
+			},
+			expected: expected{
+				err: "unexpected token '.' at index 0",
+			},
+		},
 	}
 
 	for idx, test := range tests {
@@ -76,6 +84,138 @@ func Test_ScriptEngine_Compile(t *testing.T) {
 			assert.Equal(t, test.expected.compiled, actual)
 		})
 	}
+}
+
+func Test_ScriptEngine_Evaluate(t *testing.T) {
+
+	engine := &ScriptEngine{}
+
+	type input struct {
+		root, current interface{}
+		expression    string
+		options       *option.QueryOptions
+	}
+
+	type expected struct {
+		value interface{}
+		err   string
+	}
+
+	tests := []struct {
+		input    input
+		expected expected
+	}{
+		{
+			input: input{
+				expression: "",
+			},
+			expected: expected{
+				err: "invalid expression. is empty",
+			},
+		},
+		{
+			input: input{
+				root:       "root",
+				current:    "current",
+				expression: "nil",
+			},
+			expected: expected{
+				value: nil,
+			},
+		},
+		{
+			input: input{
+				root:       "root",
+				current:    "current",
+				expression: "null",
+			},
+			expected: expected{
+				value: nil,
+			},
+		},
+		{
+			input: input{
+				root:       "root",
+				current:    "current",
+				expression: "$",
+			},
+			expected: expected{
+				value: "'root'",
+			},
+		},
+		{
+			input: input{
+				root:       "root",
+				current:    "current",
+				expression: "@",
+			},
+			expected: expected{
+				value: "'current'",
+			},
+		},
+		{
+			input: input{
+				root:       "root",
+				current:    "current",
+				expression: "other",
+			},
+			expected: expected{
+				value: "other",
+			},
+		},
+		{
+			input: input{
+				expression: "1+fish",
+			},
+			expected: expected{
+				err: "invalid argument. expected number",
+			},
+		},
+		{
+			input: input{
+				expression: "@[-1]==2",
+				current:    []interface{}{0, 2},
+			},
+			expected: expected{
+				value: true,
+			},
+		},
+		{
+			input: input{
+				expression: "@.name=~'hello.*'",
+				current: map[string]interface{}{
+					"name": "hello world",
+				},
+			},
+			expected: expected{
+				value: true,
+			},
+		},
+		{
+			input: input{
+				expression: ".@.name=~'hello.*'",
+				current: map[string]interface{}{
+					"name": "hello world",
+				},
+			},
+			expected: expected{
+				err: "unexpected token '.' at index 0",
+			},
+		},
+	}
+
+	for idx, test := range tests {
+		t.Run(fmt.Sprintf("%d", idx), func(t *testing.T) {
+			actual, err := engine.Evaluate(test.input.root, test.input.current, test.input.expression, test.input.options)
+			if test.expected.err == "" {
+				assert.Nil(t, err)
+			} else {
+				assert.EqualError(t, err, test.expected.err)
+			}
+			assert.Equal(t, test.expected.value, actual)
+		})
+	}
+
 }
 
 func Test_ScriptEngine_buildOperators(t *testing.T) {
@@ -437,6 +577,68 @@ func Test_ScriptEngine_buildOperators(t *testing.T) {
 				},
 			},
 		},
+		{
+			input: input{
+				expression: "@.key=~'hello.*'",
+				tokens:     defaultTokens,
+			},
+			expected: expected{
+				operator: &regexOperator{
+					arg1: currentKey,
+					arg2: "'hello.*'",
+				},
+			},
+		},
+		{
+			input: input{
+				expression: ".$",
+				tokens:     defaultTokens,
+			},
+			expected: expected{
+				operator: nil,
+				err:      "unexpected token '.' at index 0",
+			},
+		},
+		{
+			input: input{
+				expression: ".||",
+				tokens:     []string{"||"},
+			},
+			expected: expected{
+				operator: nil,
+				err:      "",
+			},
+		},
+		{
+			input: input{
+				expression: "||.",
+				tokens:     []string{"||"},
+			},
+			expected: expected{
+				operator: nil,
+				err:      "",
+			},
+		},
+		{
+			input: input{
+				expression: ".||.",
+				tokens:     []string{"||"},
+			},
+			expected: expected{
+				operator: &orOperator{arg1: ".", arg2: "."},
+				err:      "",
+			},
+		},
+		{
+			input: input{
+				expression: "||.",
+				tokens:     []string{"||", "&&"},
+			},
+			expected: expected{
+				operator: nil,
+				err:      "",
+			},
+		},
 	}
 
 	for idx, test := range tests {
@@ -452,16 +654,15 @@ func Test_ScriptEngine_buildOperators(t *testing.T) {
 	}
 }
 
-func Test_ScriptEngine_Evaluate(t *testing.T) {
+func Test_ScriptEngine_parseArgument(t *testing.T) {
 
 	engine := &ScriptEngine{}
 
 	type input struct {
-		root, current interface{}
-		expression    string
-		options       *option.QueryOptions
+		argument string
+		tokens   []string
+		options  *option.QueryOptions
 	}
-
 	type expected struct {
 		value interface{}
 		err   string
@@ -473,84 +674,51 @@ func Test_ScriptEngine_Evaluate(t *testing.T) {
 	}{
 		{
 			input: input{
-				expression: "",
+				argument: ".$",
+				tokens:   defaultTokens,
 			},
 			expected: expected{
-				err: "invalid expression. is empty",
+				err: "unexpected token '.' at index 0",
 			},
 		},
 		{
 			input: input{
-				root:       "root",
-				current:    "current",
-				expression: "nil",
+				argument: "1||1",
+				tokens:   defaultTokens,
 			},
 			expected: expected{
-				value: nil,
+				value: &orOperator{arg1: "1", arg2: "1"},
 			},
 		},
 		{
 			input: input{
-				root:       "root",
-				current:    "current",
-				expression: "null",
+				argument: "[1,2,3]",
+				tokens:   defaultTokens,
 			},
 			expected: expected{
-				value: nil,
+				value: []interface{}{
+					float64(1),
+					float64(2),
+					float64(3),
+				},
 			},
 		},
 		{
 			input: input{
-				root:       "root",
-				current:    "current",
-				expression: "$",
+				argument: `{"key":"value"}`,
+				tokens:   defaultTokens,
 			},
 			expected: expected{
-				value: "'root'",
-			},
-		},
-		{
-			input: input{
-				root:       "root",
-				current:    "current",
-				expression: "@",
-			},
-			expected: expected{
-				value: "'current'",
-			},
-		},
-		{
-			input: input{
-				root:       "root",
-				current:    "current",
-				expression: "other",
-			},
-			expected: expected{
-				value: "other",
-			},
-		},
-		{
-			input: input{
-				expression: "1+fish",
-			},
-			expected: expected{
-				err: "invalid argument. expected number",
-			},
-		},
-		{
-			input: input{
-				expression: "@[-1]==2",
-				current:    []interface{}{0, 2},
-			},
-			expected: expected{
-				value: true,
+				value: map[string]interface{}{
+					"key": "value",
+				},
 			},
 		},
 	}
 
 	for idx, test := range tests {
 		t.Run(fmt.Sprintf("%d", idx), func(t *testing.T) {
-			actual, err := engine.Evaluate(test.input.root, test.input.current, test.input.expression, test.input.options)
+			actual, err := engine.parseArgument(test.input.argument, test.input.tokens, test.input.options)
 			if test.expected.err == "" {
 				assert.Nil(t, err)
 			} else {
@@ -559,5 +727,4 @@ func Test_ScriptEngine_Evaluate(t *testing.T) {
 			assert.Equal(t, test.expected.value, actual)
 		})
 	}
-
 }
